added separate classes for disabling tests on system property conditions
